rydd i database-excpetion kode
