+++ src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/archive/PeriodicVarselArchiver.kt
@@ -2,7 +2,7 @@ package no.nav.personbruker.dittnav.eventaggregator.archive
 
 import no.nav.personbruker.dittnav.eventaggregator.common.LocalDateTimeHelper.nowAtUtc
 import no.nav.personbruker.dittnav.eventaggregator.common.PeriodicJob
-import no.nav.personbruker.dittnav.eventaggregator.common.exceptions.RetriableDatabaseException
+import no.nav.personbruker.dittnav.eventaggregator.common.database.RetriableDatabaseException
 import no.nav.personbruker.dittnav.eventaggregator.config.EventType
 import no.nav.personbruker.dittnav.eventaggregator.done.jobs.PeriodicDoneEventWaitingTableProcessor
 import no.nav.personbruker.dittnav.eventaggregator.varsel.VarselType
