--- src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/exceptions/AggregatorBatchUpdateException.kt
@@ -1,5 +1,0 @@
-package no.nav.personbruker.dittnav.eventaggregator.common.exceptions
-
-class AggregatorBatchUpdateException(message: String, cause: Throwable?) : RetriableDatabaseException(message, cause) {
-    constructor(message: String) : this(message, null)
-}
