--- src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/exceptions/RetriableDatabaseException.kt
@@ -1,3 +1,0 @@
-package no.nav.personbruker.dittnav.eventaggregator.common.exceptions
-
-open class RetriableDatabaseException(message: String, cause: Throwable?) : AbstractPersonbrukerException(message, cause)
