--- src/main/kotlin/no/nav/personbruker/dittnav/eventaggregator/common/exceptions/UnretriableDatabaseException.kt
@@ -1,3 +1,0 @@
-package no.nav.personbruker.dittnav.eventaggregator.common.exceptions
-
-class UnretriableDatabaseException(message: String, cause: Throwable?) : AbstractPersonbrukerException(message, cause)
